backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.115.0
uvicorn[standard]==0.32.0

# Database
sqlalchemy==2.0.36
psycopg2-binary==2.9.10

# Caching
redis==5.2.0

# Real-time communication
python-socketio==5.11.4

# Configuration
pydantic-settings==2.6.1

# AI/ML
langchain==0.3.15
langchain-openai==0.2.8
openai==1.57.3
qdrant-client==1.11.3

# HTTP client
httpx==0.28.1
